move MessageCard and TransactionThreadCard to separate files
