+++ client/src/types/index.ts
@@ -13,3 +13,17 @@ export type User = {
     id: number,
     firstName: string,
 };
+
+export type Transaction = {
+    id: number,
+    donater: User,
+    requester: User,
+    asset: Pick<Asset, 'id' | 'title'>,
+};
+
+export type Message = {
+    id: number,
+    text: string,
+    transactionId: number,
+    user: User,
+};
